feat: standalone intent service support
