+++ requirements/requirements.txt
@@ -5,6 +5,6 @@ combo-lock>=0.2.2, <0.4
 
 ovos-utils[extras]>=0.6.0,<1.0.0
 ovos_bus_client>=0.1.4,<2.0.0
-ovos-plugin-manager>=1.0.1,<2.0.0
+ovos-plugin-manager>=1.0.3,<2.0.0
 ovos-config>=0.0.13,<2.0.0
 ovos-workshop>=7.0.2,<8.0.0
